Added optional throwable parameter to exceptions
