+++ tracker/src/main/java/es/eucm/tracker/exceptions/ExtensionException.java
@@ -17,11 +17,13 @@
 // Translated by CS2J (http://www.cs2j.com): 05/11/2018 15:29:15
 //
 
-package es.eucm.tracker.Exceptions;
+package es.eucm.tracker.exceptions;
 
-public class ExtensionXApiException extends XApiException {
-	public ExtensionXApiException(String message) {
+public class ExtensionException extends TrackerException {
+	public ExtensionException(String message) {
 		super(message);
 	}
-
+	public ExtensionException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
