+++ tracker/src/main/java/es/eucm/tracker/exceptions/TargetXApiException.java
@@ -17,11 +17,13 @@
 // Translated by CS2J (http://www.cs2j.com): 05/11/2018 15:29:15
 //
 
-package es.eucm.tracker.Exceptions;
+package es.eucm.tracker.exceptions;
 
 public class TargetXApiException extends XApiException {
 	public TargetXApiException(String message) {
 		super(message);
 	}
-
+	public TargetXApiException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
