+++ tracker/src/main/java/es/eucm/tracker/exceptions/TraceException.java
@@ -17,11 +17,13 @@
 // Translated by CS2J (http://www.cs2j.com): 05/11/2018 15:29:15
 //
 
-package es.eucm.tracker.Exceptions;
+package es.eucm.tracker.exceptions;
 
-public class KeyExtensionException extends ExtensionException {
-	public KeyExtensionException(String message) {
+public class TraceException extends TrackerException {
+	public TraceException(String message) {
 		super(message);
 	}
-
+	public TraceException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
