+++ tracker/src/main/java/es/eucm/tracker/exceptions/TrackerException.java
@@ -17,11 +17,15 @@
 // Translated by CS2J (http://www.cs2j.com): 05/11/2018 15:29:15
 //
 
-package es.eucm.tracker.Exceptions;
+package es.eucm.tracker.exceptions;
 
-public class ValueExtensionException extends ExtensionException {
-	public ValueExtensionException(String message) {
+public class TrackerException extends Exception {
+
+	public TrackerException(String message) {
 		super(message);
 	}
 
+	public TrackerException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
