+++ tracker/src/main/java/es/eucm/tracker/exceptions/ValueExtensionException.java
@@ -17,11 +17,13 @@
 // Translated by CS2J (http://www.cs2j.com): 05/11/2018 15:29:15
 //
 
-package es.eucm.tracker.Exceptions;
+package es.eucm.tracker.exceptions;
 
-public class ActorXApiException extends XApiException {
-	public ActorXApiException(String message) {
+public class ValueExtensionException extends ExtensionException {
+	public ValueExtensionException(String message) {
 		super(message);
 	}
-
+	public ValueExtensionException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
