+++ tracker/src/main/java/es/eucm/tracker/exceptions/VerbXApiException.java
@@ -17,11 +17,13 @@
 // Translated by CS2J (http://www.cs2j.com): 05/11/2018 15:29:15
 //
 
-package es.eucm.tracker.Exceptions;
+package es.eucm.tracker.exceptions;
 
-public class XApiException extends TrackerException {
-	public XApiException(String message) {
+public class VerbXApiException extends XApiException {
+	public VerbXApiException(String message) {
 		super(message);
 	}
-
+	public VerbXApiException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
