+++ tracker/src/main/java/es/eucm/tracker/exceptions/XApiException.java
@@ -17,11 +17,13 @@
 // Translated by CS2J (http://www.cs2j.com): 05/11/2018 15:29:15
 //
 
-package es.eucm.tracker.Exceptions;
+package es.eucm.tracker.exceptions;
 
-public class TrackerException extends Exception {
-	public TrackerException(String message) {
+public class XApiException extends TrackerException {
+	public XApiException(String message) {
 		super(message);
 	}
-
+	public XApiException(String message, Throwable cause) {
+		super(message, cause);
+	}
 }
